+++ src/brasil/gov/portlets/portlets/audiogallery.py
@@ -19,24 +19,31 @@ from zope.interface import implements
 
 
 class IAudioGalleryPortlet(IPortletDataProvider):
-    '''Portal Padrão: Portlet de galeria de áudio.
+    '''Portal Padrao: Audio Gallery Portlet.
     '''
 
     show_header = schema.Bool(
-        title=_(u'Mostrar título'),
-        description=_(u'Se habilitado mostra o título.'),
+        title=_(u'show_title',
+                default=u'Show title'),
+        description=_(u'show_title_description',
+                      default=u'If enabled, shows the title.'),
         required=True,
         default=False)
 
     header = schema.TextLine(
-        title=_(u'Texto do título'),
-        description=_(u'Texto do título do portlet.'),
+        title=_(u'title_text',
+                default=u'Title text'),
+        description=_(u'title_text_description',
+                      default=u'Portlet text of the title.'),
         required=True,
-        default=_(u'Portal Padrão Galeria de Áudios'))
+        default=_(u'title_portlet_audiogallery',
+                  default=u'Portal Padrao Audio Gallery'))
 
     header_type = schema.Choice(
-        title=_(u'Tipo de título'),
-        description=_(u'Tipo de título que será exibido.'),
+        title=_(u'title_type',
+                default=u'Title type'),
+        description=_(u'title_type_description',
+                      default=u'Title type that will be shown.'),
         values=(u'H1',
                 u'H2',
                 u'H3',
@@ -46,31 +53,42 @@ class IAudioGalleryPortlet(IPortletDataProvider):
     )
 
     show_footer = schema.Bool(
-        title=_(u'Mostrar rodapé'),
-        description=_(u'Se habilitado mostra o rodapé.'),
+        title=_(u'show_footer',
+                default=u'Show footer'),
+        description=_(u'show_footer_description',
+                      default=u'If enabled, shows the footer.'),
         required=True,
         default=False)
 
     footer = schema.TextLine(
-        title=_(u'Texto do rodapé'),
-        description=_(u'Texto do rodapé do portlet.'),
+        title=_(u'footer_text',
+                default=u'Footer text'),
+        description=_(u'footer_text_description',
+                      default=u'Portlet footer text.'),
         required=False)
 
     footer_url = schema.TextLine(
-        title=_(u'Link do rodapé'),
-        description=_(u'Link do rodapé do portlet.'),
+        title=_(u'footer_url',
+                default=u'Footer URL'),
+        description=_(u'footer_url_description',
+                      default=u'Portlet footer URL.'),
         required=False)
 
     limit = schema.Int(
-        title=_(u'Quantidade de itens a exibir'),
-        description=_(u'Informe o total de itens que devem ser exibidos no '
-                      u'portlet.'),
+        title=_(u'limit',
+                default=u'Number of items to show'),
+        description=_(u'limit_description',
+                      default=u'Total itens that should be displayed in ' +
+                              u'the portlet.'),
         required=True,
         default=5)
 
     collection = schema.Choice(
-        title=_(u'Coleção'),
-        description=_(u'Pesquisa a coleção utilizada no portlet.'),
+        title=_(u'collection',
+                default=u'Collection'),
+        description=_(u'collection_description',
+                      default=u'Searchs the collection that will be used ' +
+                              u'in the portlet.'),
         required=True,
         source=SearchableTextSourceBinder(
             {'portal_type': ('Topic', 'Collection')},
@@ -82,7 +100,8 @@ class Assignment(base.Assignment):
     implements(IAudioGalleryPortlet)
 
     show_header = False
-    header = _(u'Portal Padrão Galeria de Áudios')
+    header = _(u'title_portlet_audiogallery',
+               default=u'Portal Padrao Audio Gallery')
     header_type = u'H2'
     show_footer = False
     footer = u''
@@ -92,7 +111,8 @@ class Assignment(base.Assignment):
 
     def __init__(self,
                  show_header=False,
-                 header=_(u'Portal Padrão Galeria de Áudios'),
+                 header=_(u'title_portlet_audiogallery',
+                          default=u'Portal Padrao Audio Gallery'),
                  header_type=u'H2',
                  show_footer=False,
                  footer=u'',
@@ -201,8 +221,9 @@ class AddForm(base.AddForm):
     form_fields = form.Fields(IAudioGalleryPortlet)
     form_fields['collection'].custom_widget = UberSelectionWidget
 
-    label = _(u'Adicionar Portlet Portal Padrão Galeria de Áudios')
-    description = _(u'Este portlet mostra uma Galeria de Áudios.')
+    label = _(u'Add Portlet Portal Padrao Audio Gallery')
+    description = _(u'audiogallery_portlet_description',
+                    default=u'This portlet shows an Audio Gallery.')
 
     def create(self, data):
         return Assignment(**data)
@@ -213,5 +234,6 @@ class EditForm(base.EditForm):
     form_fields = form.Fields(IAudioGalleryPortlet)
     form_fields['collection'].custom_widget = UberSelectionWidget
 
-    label = _(u'Editar Portlet Portal Padrão Galeria de Áudios')
-    description = _(u'Este portlet mostra uma Galeria de Áudios.')
+    label = _(u'Edit Portlet Portal Padrao Audio Gallery')
+    description = _(u'audiogallery_portlet_description',
+                    default=u'This portlet shows an Audio Gallery.')